# Simplify AVX usage