Follow up on @ReneHollander's PR #262 

Took a quick stab at simplifying AVX usage. Almost all of it was various types of dot products, so I added a function `template<T> dot(const T* x, const T* y, int f)` – if `USE_AVX` is defined then I have a template specialization for T = float that implements it using AVX