@erikbern 2018-01-19
Ok I added a new primitive dot3 to compute the products xx, yy, and xy at the same time now. Think it should do the trick!