@erikbern 2018-01-19
Unscientific benchmark:

After this PR:
Building + searching: 246.822s
Searching (best of 3): 8.584s

Current master:
Building + searching: 232.608s
Searching (best of 3): 8.243s

Master from before the AVX instructions
Building + searching: 230.247s
Searching (best of 3): 8.505s

This is really weird. No difference? I know for a fact that my laptop uses AVX. Maybe it doesn't  give any speedup on my CPU (?)