@erikbern 2018-01-19
seems like my processor might just be too slow: https://stackoverflow.com/questions/43245787/vectorization-speed-up-expected-for-sse-avx-and-avx2

i'll try on a different CPU tomorrow