@erikbern 2018-01-19
rebased on master (#266 in particular)