@ReneHollander 2018-01-19
I checked the assembly and without declaring `get_norm` and `normalize` like you did with `dot3` and `dot`, g++ won't use it. The float implementation of `two_means` forced the usage of the float variants and thus created the speedup. My optimized `two_means` is mostly unreadable and does not affect performance at all.
```diff
Index: src/annoylib.h
IDEA additional info:
Subsystem: com.intellij.openapi.diff.impl.patch.CharsetEP
<+>UTF-8
===================================================================
--- src/annoylib.h	(revision 35be9a1b8ff8c0f7ccfece330b1386691993c0ae)
+++ src/annoylib.h	(date 1516371870000)
@@ -237,7 +237,8 @@
   }
 }
 
-static inline float get_norm(float *v, int f) {
+template<>
+inline float get_norm<float>(float *v, int f) {
   float sq_norm = 0;
   int i = f;
   if (f > 7) {
@@ -258,7 +259,8 @@
   return sqrt(sq_norm);
 }
 
-static inline void normalize(float *v, int f) {
+template<>
+inline void normalize<float>(float *v, int f) {
   float norm = get_norm(v, f);
   __m256 v_norm = _mm256_set1_ps(norm);
 

```
I also ran my simple benchmarks, and I still see the same results I posted in my other PR. Index build time gets halfed, Query time reduced by 25% or something like that.