@erikbern 2018-01-19
Ok great – I'm glad my rusty C++ intuition was correct and template specialization is the way to go here.

I'll probably merge this for now