@ReneHollander 2018-01-19
I did some more benchmarking (all synthetic data):
There is a noticable improvement on my i7 4770K in WSL on Windows 10. At no point is the AVX optimized code slower, and it performs the best if the number if dimensions is a multiple of 8 (as expected). I think I set the k for the search test a bit too low, but I did not want to rerun the test (I could not use my PC for 6 hours...). I hope you see similar improvements in your tests!
# Euclidian
![euclidian_build_time](https://user-images.githubusercontent.com/5379596/35174506-f28e913c-fd6f-11e7-8e4d-34f941865c22.png)
![euclidian_search_time](https://user-images.githubusercontent.com/5379596/35174508-f528aa22-fd6f-11e7-8a80-0410758be278.png)
(Search time in us)
# Angular
![angular_build_time](https://user-images.githubusercontent.com/5379596/35174513-fb9c516a-fd6f-11e7-80e7-0d4e8b7db3d0.png)
![angular_search_time](https://user-images.githubusercontent.com/5379596/35174514-fcff0bba-fd6f-11e7-9fc5-f18bdd2ce858.png)