@erikbern 2018-01-20
Awesome! Was actually just running some benchmarks myself, this time on a c4.2xlarge:

current master:
glove-25 build+search: 136.447s
glove-25 search (best of 3): 5.697s
fashion-mnist build+search: 19.126s
fashion-mnist search (best of 3): 8.798s

pre-avx:
glove-25 build+search: 152.749s
glove-25 search (best of 3): 6.351s
fashion-mnist build+search: 33.285s
fashion-mnist search (best of 3): 18.096s

this branch:
glove-25 build+search: 146.448s
glove-25 search (best of 3): 5.754s
fashion-mnist build+search: 20.259s
fashion-mnist search (best of 3): 9.163s

looks like master and this branch is basically same (great, my refactoring didn't ruin anything) and much faster than pre-avx, by up to a 2x factor. very nice @ReneHollander !